pyproject.toml
[tool.poetry]
name = "talk-pythonie-meetup-kafka"
version = "0.1.0"
description = ""
authors = ["Stéphane Wirtel <stephane@wirtel.be>"]

[tool.poetry.dependencies]
python = "^3.8"
faust = "^1.10.4"
minio = "^5.0.10"
uvicorn = "^0.11.5"
ujson = "^3.0.0"
mode = "^4.3.2"
fastapi = "^0.55.1"
httpx = "^0.13.3"
bs4 = "^0.0.1"
splinter = "^0.13.0"
pillow = "^7.1.2"
aiosmtplib = "^1.1.3"
python-decouple = "^3.3"
pydantic = {version = "^1.5.1", extras = ["dotenv", "email"]}
yarl = "^1.4.2"
flake8 = "^3.8.3"

[tool.poetry.dev-dependencies]
pre-commit = "^2.5.0"
mypy = "^0.780"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
